Split selectors into several files
