+++ packages/redux-saga-requests/src/selectors/get-query.spec.js
@@ -1,4 +1,4 @@
-import { getQuery } from './selectors';
+import getQuery from './get-query';
 
 describe('selectors', () => {
   describe('getQuery', () => {
