+++ packages/redux-saga-requests/src/selectors/index.js
@@ -1,0 +1,2 @@
+export { default as getQuery } from './get-query';
+export { default as getMutation } from './get-mutation';
